fix: show last line of files in diff viewer
The diff viewer was subtracting an extra buffer line when calculating visible height, preventing the last line of files from being displayed. Updated the calculation to remove the unnecessary -1 buffer, changing from -6 to -5 to account only for:
- 1 for app header
- 1 for panel header
- 2 for file header
- 1 for status bar

This allows the full available space to be used for displaying file content.